Update windows-sys crate version to 0.59 (#886)
- Updates window.sys to version 0.58.
- A key change to this version is that the HANDLE type is now
`*mut c_void`, so the code was changed to accomodate this.
- An important change is the manual implementation of the `Send`
trait for `ChildExistWatcher`. The trait can no longer derived
automatically because `*mut c_void` is not Send. However, pointers
are not Send more as a lint to developers (eg: a pointer to Rc<T>
would be safe), but `c_void` is Send and therefore it should be
fine to derive it manually for the pointer.

See https://doc.rust-lang.org/nomicon/send-and-sync.html for details.

+++ teletypewriter/src/windows/pipes.rs
@@ -201,7 +201,7 @@ impl Drop for EventedAnonRead {
 
         // Stop reader thread waiting for pipe contents
         unsafe {
-            CancelSynchronousIo(thread.as_raw_handle() as isize);
+            CancelSynchronousIo(thread.as_raw_handle());
         }
 
         thread
